Cache attributes within transaction (#5417)
## What is the goal of this PR?
In situations when we insert multiple attributes within a transaction, especially when the attribute value is a super node (for example `male/female` gender), it is beneficial to cache the inserted attribute instead of doing a db lookup which requires an index check. Otherwise we need to fetch the same attribute multiple times. This PR introduces an extra transaction-based attribute cache to allow to cache attributes. 

## What are the changes implemented in this PR?
Introduce attribute cache in `TransactionCache` to store attributes created in that transaction.